Better handling of lua arguments for basic types
Using `LuaArg<T>(L, index)` for getting a index-th argument as type T.
Check if the argument is of the desired type and throw an error if not.

Using `LuaArgOptional<T>(L, index, default_value)` to obtain value of an
optional argument. If argument is not specified, use `default_value` as
its value.