Removal of Python plotting subroutine from SLDFE Quadrature Routine
This removes the python subroutine for plotting and displaying angular discretization from the sldfe_sq_quadrature.cc and places into a new directory in tools. The file format for the plotting data was changed to a csv file with a header for readability, and the relevant test case for this routine was updated accordingly.
Additionally, the README file in tools was updated to include a description of how the new python plotting tool should be used.