Disable clang-tidy linting in Varying data type
`Varying` seems like something that std::variant does, but maybe the older C++
standard used during development of this code did not have std::variant and that's
why this exists. Simple fixes do not seem to be possible, this type is tied
with `InputParameters` and was heavily utilized by the lua layer which is now gone.
However simple removal is not possible either.
The code relies on this capability, and the possible portalibility issue should
not be a a problem untill users build opensn with clang and their application
with a different compiler.